User input using argparse, implement PEP guidelines
