+++ To-do-list/manan/.gitignore
@@ -1,0 +1,5 @@
+*.png
+*.json
+*.txt
+__pycache__
+
