+++ To-do-list/manan/readme.md
@@ -1,0 +1,17 @@
+# **T0-DO LIST**
+---
+###### A simple command line to-do list for linux
+#
+---
+#### Requirements
+* python 3+
+* pip
+```sudo apt-get install pip```
+---
+#### Installation
+* ```pip install python-vlc```
+* ```pip install Pillow```
+* ```pip install Pygments```
+OR
+* ```pip install -r requirements.txt```
+
