+++ To-do-list/manan/requirements.txt
@@ -1,0 +1,4 @@
+Pygments==2.2.0
+Pillow==3.1.0
+python-vlc==3.0.102
+
